% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM_main.R
\name{DualALM_main}
\alias{DualALM_main}
\title{This function contains the main loop for Augmented Lagrange Multipler}
\usage{
DualALM_main(LL, parmain, x, y, u, v)
}
\description{
This function contains the main loop for Augmented Lagrange Multipler
}
