% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM for solving the NPMLE with unknown supports
input: observations X \in R^{n\emph{d}
SIGMA: n numbers of diagonal covararince matrices \in R^{d}n}
m: number of grid points
option$grid_initial: initial supports}
\usage{
EM(X, SIGMA, m, options)
}
\description{
EM for solving the NPMLE with unknown supports
input: observations X \in R^{n\emph{d}
SIGMA: n numbers of diagonal covararince matrices \in R^{d}n}
m: number of grid points
option$grid_initial: initial supports
}
