% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observation.R
\name{generate_observation}
\alias{generate_observation}
\title{generate observations in R^2
Input: n # observations
fig_option = 1 two concentric circles
= 2 triangle
= 3 digit 8
= 4 letter A
= 5 circle of radius 6
= 6 theta_i = 0
= 7 theta_i = 0, 6\emph{e1, or 6}e2
= 8 theta_i = N(0,SIGMA);
= 9: theta_i has 6 support points \link{0  0 0 ... 0} \link{1  0 0 ... 0} \link{-1 0 0 ... 0}
\link{0  1 0 ... 0} \link{1  1 0 ... 0} \link{-1 1 0 ... 0}
sigma_option = 1, I2,
= 2, SIGMA_i = diag(ai,bi),ai,bi ~ Uniform\link{1,3}
Output: observations X \in R^{n\emph{d} (n observations in each row)
theta \in R^{n}d} true signal
SIGMA \in R^{d*n} if sigma_option = 2, SIGMA(:,:,i) = \link{ai,bi}}
\usage{
generate_observation(n, fig_option, sigma_option, d)
}
\description{
generate observations in R^2
Input: n # observations
fig_option = 1 two concentric circles
= 2 triangle
= 3 digit 8
= 4 letter A
= 5 circle of radius 6
= 6 theta_i = 0
= 7 theta_i = 0, 6\emph{e1, or 6}e2
= 8 theta_i = N(0,SIGMA);
= 9: theta_i has 6 support points \link{0  0 0 ... 0} \link{1  0 0 ... 0} \link{-1 0 0 ... 0}
\link{0  1 0 ... 0} \link{1  1 0 ... 0} \link{-1 1 0 ... 0}
sigma_option = 1, I2,
= 2, SIGMA_i = diag(ai,bi),ai,bi ~ Uniform\link{1,3}
Output: observations X \in R^{n\emph{d} (n observations in each row)
theta \in R^{n}d} true signal
SIGMA \in R^{d*n} if sigma_option = 2, SIGMA(:,:,i) = \link{ai,bi}
}
