% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqmr.R
\name{psqmr}
\alias{psqmr}
\title{psqmr:  preconditioned symmetric QMR with left (symmetric) preconditioner.
b = rhs vector.
resnrm = norm of qmr-generated residual vector b-Ax.}
\usage{
psqmr(matvecfname, b, par, x0, Ax0)
}
\description{
psqmr:  preconditioned symmetric QMR with left (symmetric) preconditioner.
b = rhs vector.
resnrm = norm of qmr-generated residual vector b-Ax.
}
